refactor: split protocol mod in multiple mods
