+++ src/lib.rs
@@ -27,7 +27,7 @@
 //! }
 //! ```
 
-use crate::protocol::{Status, SystemInfo};
+use crate::protocol::SystemInfo;
 use serde::Deserialize;
 use std::net::SocketAddr;
 use thiserror::Error;
@@ -63,6 +63,13 @@ pub enum Error {
     Network(#[from] tokio::io::Error),
     #[error("Failed to parse xml: {0}")]
     Xml(#[from] serde_xml_rs::Error),
+}
+
+#[derive(Debug, Deserialize, Clone, PartialEq, Eq)]
+pub struct Status {
+    pub code: i32,
+    #[serde(rename = "$value")]
+    pub text: String,
 }
 
 impl Plugin {
