+++ src/protocol/mod.rs
@@ -1,0 +1,9 @@
+mod serialize;
+
+#[cfg(feature = "simulator-time")]
+mod simulator_time;
+mod system_info;
+
+#[cfg(feature = "simulator-time")]
+pub(crate) use simulator_time::*;
+pub use system_info::*;
