+++ src/protocol/serialize/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "simulator-time")]
+mod time_milli_seconds;
+
+#[cfg(feature = "simulator-time")]
+pub use time_milli_seconds::*;
